requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.53.0
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.41.1
matplotlib==3.7.5
networkx==3.1
numba==0.58.1
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
tifffile==2023.7.10
torch==1.10.1+cu111
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.19.2
